fix locations selector
